move switch component to ui folder
